refac: eliminated redundant call for nvim
The Neovim version of `CursorMoved` triggers when the cursor enters a window/tab,
whether by moving, or by closing.  So `source_move` is called twice in such a situation.

However, the Vim version of `CursorMoved` does not trigger on window/tab enter. So `source_move`
is called once as needed.

This change eliminates the call for Nvim but not Vim.  It also changes occurences of `has(...)`
to script-wide variables.